+++ plugins/module_utils/mh/module_helper.py
@@ -6,19 +6,16 @@
 from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
-from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.common.dict_transformations import dict_merge
 
-from ansible_collections.community.general.plugins.module_utils.mh.base import ModuleHelperBase
+from ansible_collections.community.general.plugins.module_utils.mh.base import ModuleHelperBase, AnsibleModule
 from ansible_collections.community.general.plugins.module_utils.mh.mixins.cmd import CmdMixin
 from ansible_collections.community.general.plugins.module_utils.mh.mixins.state import StateMixin
 from ansible_collections.community.general.plugins.module_utils.mh.mixins.deps import DependencyMixin
-from ansible_collections.community.general.plugins.module_utils.mh.deco import module_fails_on_exception
-from ansible_collections.community.general.plugins.module_utils.mh.mixins.vars import VarDict
+from ansible_collections.community.general.plugins.module_utils.mh.mixins.vars import VarsMixin
 
 
-class ModuleHelper(DependencyMixin, ModuleHelperBase):
-
+class ModuleHelper(VarsMixin, DependencyMixin, ModuleHelperBase):
     _output_conflict_list = ('msg', 'exception', 'output', 'vars', 'changed')
     facts_name = None
     output_params = ()
@@ -27,15 +24,7 @@ class ModuleHelper(DependencyMixin, ModuleHelperBase):
     facts_params = ()
 
     def __init__(self, module=None):
-        self.vars = VarDict()
-        self._changed = False
-
-        if module:
-            self.module = module
-
-        if not isinstance(self.module, AnsibleModule):
-            self.module = AnsibleModule(**self.module)
-
+        super(ModuleHelper, self).__init__(module)
         for name, value in self.module.params.items():
             self.vars.set(
                 name, value,
@@ -45,12 +34,6 @@ class ModuleHelper(DependencyMixin, ModuleHelperBase):
                 fact=name in self.facts_params,
             )
 
-    def update_vars(self, meta=None, **kwargs):
-        if meta is None:
-            meta = {}
-        for k, v in kwargs.items():
-            self.vars.set(k, v, **meta)
-
     def update_output(self, **kwargs):
         self.update_vars(meta={"output": True}, **kwargs)
 
@@ -59,14 +42,6 @@ class ModuleHelper(DependencyMixin, ModuleHelperBase):
 
     def _vars_changed(self):
         return any(self.vars.has_changed(v) for v in self.vars.change_vars())
-
-    @property
-    def changed(self):
-        return self._changed
-
-    @changed.setter
-    def changed(self, value):
-        self._changed = value
 
     def has_changed(self):
         return self.changed or self._vars_changed()
@@ -88,15 +63,6 @@ class ModuleHelper(DependencyMixin, ModuleHelperBase):
                 result["_" + varname] = result[varname]
                 del result[varname]
         return result
-
-    @module_fails_on_exception
-    def run(self):
-        self.fail_on_missing_deps()
-        self.__init_module__()
-        self.__run__()
-        self.__quit_module__()
-        self.module.exit_json(changed=self.has_changed(), **self.output)
-
 
 class StateModuleHelper(StateMixin, ModuleHelper):
     pass
